+++ tests/testQuadraticGenerator.cpp
@@ -9,13 +9,13 @@
 
 int main()
 {
-  int nr_points = 1001;
-  int extra_points = 200;
+  unsigned int nr_points = 1001;
+  unsigned int extra_points = 200;
   mc_rbdyn::QuadraticGenerator generator(1., 5., nr_points);
   std::vector<std::pair<double, double>> vec;
   vec.reserve(nr_points + extra_points);
   double speed, pos;
-  for(int i = 0; i < nr_points + extra_points; ++i)
+  for(size_t i = 0; i < nr_points + extra_points; ++i)
   {
     generator.next(pos, speed);
     vec.push_back(std::make_pair(speed, pos));
